+++ pkgs/native_assets_builder/test/build_runner/version_skew_test.dart
@@ -1,0 +1,36 @@
+// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:test/test.dart';
+
+import '../helpers.dart';
+import 'helpers.dart';
+
+const Timeout longTimeout = Timeout(Duration(minutes: 5));
+
+void main() async {
+  for (final version in ['v1_0_0', 'v1_5_0']) {
+    test('native_add build $version', timeout: longTimeout, () async {
+      await inTempDir((tempUri) async {
+        await copyTestProjects(targetUri: tempUri);
+        final packageUri = tempUri.resolve('native_add_$version/');
+
+        // First, run `pub get`, we need pub to resolve our dependencies.
+        await runPubGet(
+          workingDirectory: packageUri,
+          logger: logger,
+        );
+
+        {
+          final result = await build(
+            packageUri,
+            logger,
+            dartExecutable,
+          );
+          expect(result.assets.length, 1);
+        }
+      });
+    });
+  }
+}